+++ src/Controller/Admin/PagesController.php
@@ -125,7 +125,7 @@ class PagesController extends AppController
             if ($this->Pages->save($page)) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect($this->referer(['action' => 'index']));
+                return $this->redirectMatchingReferer(['action' => 'index']);
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -157,7 +157,7 @@ class PagesController extends AppController
             if ($this->Pages->save($page)) {
                 $this->Flash->success(I18N_OPERATION_OK);
 
-                return $this->redirect($this->referer(['action' => 'index']));
+                return $this->redirectMatchingReferer(['action' => 'index']);
             }
 
             $this->Flash->error(I18N_OPERATION_NOT_OK);
@@ -179,6 +179,6 @@ class PagesController extends AppController
         $this->Pages->deleteOrFail($this->Pages->get($id));
         $this->Flash->success(I18N_OPERATION_OK);
 
-        return $this->redirect($this->referer(['action' => 'index']));
+        return $this->redirectMatchingReferer(['action' => 'index']);
     }
 }